Added fieldline tracer and changed namespace.
namespace fieldline is renamed to Fieldline to prevent collisions with
fieldline::core::fieldline and fieldline::trace::fieldline.

fieldline::core::fieldline holds the positions on the fieldline and the
magnetic field on the fieldline.

fieldline::core::stopCriterion is a base class to be extended for
different stopping criteria.

fieldline::trace::stopCriterionSteps is defined to stop the
field line trace after N steps.