Added routines for target representation
Fieldline::core::point represents a point in R,z with the information if
it was hit or not.

Fieldline::core::line represents a line in R,z. Two lines can be checked
for intersection and Fieldline::core::point is returned containing the
point and the information if the hit point is within the lines.

Fieldline::core::target contains the polygon in R,z representing the
target. It can be read and saved to file.