Fixed scope axiSymmetric and added setup.py
The scope Fieldline.core and Fieldline.axiSymmetric are now sorted correcly in python.

setup.py is the setup file for standard installations. It was tested on kubuntu 14.10 with boost installed.
If anyone has a different distribution please test this file. If different distros are desired with different settings multiple setup files will have to be generated.