First usable python version.
The basic field line tracing is now implemented. A magnetic equilibrium
can be loaded, the stop criterion can be defined. The field line can be
calculated.

The calculation of the field line length is now correct. Test at the
separatrix show the expected decrease of connection length away from the
separatrix.