Simplified  x-and o-point detection.
Fieldline::axiSymmetric::magneticField::get_singularity is now the
combination of the prior get_singularity and get_secondDerivation. The
runs recursively.

get_secondDerivation is obsolete and has been removed / is now part of
get_singularity.

The argument list of get_singularity changed to the R,z,phi coordinates.
In the current implementation phi is a dummy variable and is only
intended to define the argument list for non axial symmetric cases.